refactor: :recycle: move config files to `/src/configs`
