+++ .gitignore
@@ -1,8 +1,7 @@
 node_modules
 exports
 
-/src/db/config.ts
-/src/utils/templates.ts
+/src/configs
 
 *.db
 *.local/**
